+++ vec3.h
@@ -4,13 +4,13 @@
 #include <cmath>
 class Vec3 {
   public:
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 
     Vec3() { x = y = z = 0; };
-    Vec3(float _x) { x = y = z = _x; };
-    Vec3(float _x, float _y, float _z) : x(_x), y(_y), z(_z) {};
+    Vec3(double _x) { x = y = z = _x; };
+    Vec3(double _x, double _y, double _z) : x(_x), y(_y), z(_z) {};
 
     Vec3 operator-() const {
       return Vec3(-x, -y, -z);
@@ -24,50 +24,50 @@ class Vec3 {
       return *this;
     };
 
-    float length() const {
+    double length() const {
       return std::sqrt(x*x + y*y + z*z);
     };
-    float length2() const {
+    double length2() const {
       return x*x + y*y + z*z;
     };
 };
 inline Vec3 operator+(const Vec3& v1, const Vec3& v2) {
   return Vec3(v1.x + v2.x, v1.y + v2.y, v1.z + v2.z);
 }
-inline Vec3 operator+(const Vec3& v1, float k) {
+inline Vec3 operator+(const Vec3& v1, double k) {
   return Vec3(v1.x + k, v1.y + k, v1.z + k);
 }
-inline Vec3 operator+(float k, const Vec3& v2) {
+inline Vec3 operator+(double k, const Vec3& v2) {
   return Vec3(k + v2.x, k + v2.y, k + v2.z);
 }
 
 inline Vec3 operator-(const Vec3& v1, const Vec3& v2) {
   return Vec3(v1.x - v2.x, v1.y - v2.y, v1.z - v2.z);
 }
-inline Vec3 operator-(const Vec3& v1, float k) {
+inline Vec3 operator-(const Vec3& v1, double k) {
   return Vec3(v1.x - k, v1.y - k, v1.z - k);
 }
-inline Vec3 operator-(float k, const Vec3& v2) {
+inline Vec3 operator-(double k, const Vec3& v2) {
   return Vec3(k - v2.x, k - v2.y, k - v2.z);
   }
 
 inline Vec3 operator*(const Vec3 v1, const Vec3& v2) {
   return Vec3(v1.x * v2.x, v1.y * v2.y, v1.z * v2.z);
 }
-inline Vec3 operator*(const Vec3& v1, float k) {
+inline Vec3 operator*(const Vec3& v1, double k) {
   return Vec3(v1.x * k, v1.y * k, v1.z * k);
 }
-inline Vec3 operator*(float k, const Vec3& v2) {
+inline Vec3 operator*(double k, const Vec3& v2) {
   return Vec3(k * v2.x, k * v2.y, k * v2.z);
 }
 
 inline Vec3 operator/(const Vec3& v1, const Vec3& v2) {
   return Vec3(v1.x / v2.x, v1.y / v2.y, v1.z / v2.z);
 }
-inline Vec3 operator/(const Vec3& v1, float k) {
+inline Vec3 operator/(const Vec3& v1, double k) {
   return Vec3(v1.x / k, v1.y / k, v1.z / k);
 }
-inline Vec3 operator/(float k, const Vec3& v2) {
+inline Vec3 operator/(double k, const Vec3& v2) {
   return Vec3(k / v2.x, k / v2.y, k / v2.z);
 }
 
@@ -78,7 +78,7 @@ inline std::ostream& operator<<(std::ostream& stream, const Vec3& v) {
 }
 
 
-inline float dot(const Vec3& v1, const Vec3& v2) {
+inline double dot(const Vec3& v1, const Vec3& v2) {
   return v1.x*v2.x + v1.y*v2.y + v1.z*v2.z;
 }
 inline Vec3 cross(const Vec3& v1, const Vec3& v2) {
@@ -91,7 +91,7 @@ inline Vec3 normalize(const Vec3& v) {
 }
 
 
-inline Vec3 pow(const Vec3& v, float n) {
+inline Vec3 pow(const Vec3& v, double n) {
   return Vec3(std::pow(v.x, n), std::pow(v.y, n), std::pow(v.z, n));
 }
 inline Vec3 exp(const Vec3& v) {
